% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cluster.R
\name{conn_clust_rec}
\alias{conn_clust_rec}
\title{Cluster by connectivity, greedy}
\usage{
conn_clust_rec(conn_mat, stop_all = TRUE, break_ties = "random",
  min_sites = 100, max_primary = Inf)
}
\arguments{
\item{conn_mat}{binary site by site connectivity matrix, TRUE/1 if sites i and j are connected}

\item{stop_all}{Should sites with no neighbours be returned as cluster centers? If TRUE, yes, if FALSE, no.}

\item{break_ties}{How should ties be broken when multiple sites tie for having the most connected neighbours. One of "random", "first"}

\item{min_sites}{If the site with the most neighbours has less than min_sites neighbours, do not return any more cluster centers}

\item{max_primary}{Maximum number of clusters to allow.}
}
\value{
vector of site row indicies that are cluster centers.
}
\description{
Generates a clustering from a binary connectivity matrix.
}
\details{
The first cluster center is the site with the most connected neighbours.
The center and all center neighbours are the first cluster.
The second cluster center is the site with the most connected neighbours excluding any sites in the first cluster.
Each new cluster is formed from sites not yet connected to any center.
}
\examples{
Membership in each cluster is then given by conn_mat[, centers]

binary_membership <- (rowSums((membership*1) * matrix(2^(seq(ncol(membership), 1)), nrow = nrow(membership), ncol = ncol(membership), byrow = TRUE)))
binary_membership_log <- log2(binary_membership)
}
