% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{convert_csiro_cpr_copepod}
\alias{convert_csiro_cpr_copepod}
\title{Convert CSIRO CPR copepod CSV into R format}
\usage{

  convert_csiro_cpr_copepod(inFile = "/vmshare/phd/data/CSIRO_cpr/cpr_copepods_matrix.csv",
  outFormat = "archivist", outDir = "results", outFile = NULL,
  outFileSpecies = NULL)
}
\arguments{
\item{inFile}{Pathe to the original CSV file. Default "/vmshare/phd/data/CSIRO_cpr/cpr_copepods_matrix.csv"}

\item{outFormat}{Type of output. Values are ("csv", "rds", "archivist"). Default "archivist"}

\item{outDir}{Directory of output file. Defaults to "results"}

\item{outFile}{Name of output file. Defaults to inFile basename}

\item{outFileSpecies}{Name of species list. Defaults to paste0(outFile, "_species")}
}
\value{
File name as char string for outFormats "csv" and "rds" and hash string for ("archivist")
}
\description{
Takes the CSIRO CPR copepod data in CSV form and converts it into
a form suitable for processing with Gradient Forest, then
exports to different formats. The species list is exported
to the same format.
}
\details{
The magic numbers in this function are:
column indexes and names for non-Species columns
The date parsing string
}
