% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{download_seaaroundus_bio}
\alias{download_seaaroundus_bio}
\title{helper function to get SeaAroundUs cells}
\usage{

  download_seaaroundus_bio(data_dir = "/vmshare/phd/data/SeaAroundUs/cache",
  bounding_wkt = NULL, chunk_size = 100, years = 2010)
}
\arguments{
\item{data_dir}{Directory to cache SeaAroundUs data}

\item{chunk_size}{Integer indicating how many cells to fetch in a single call. Too small will slow down the download process, too large will cause server timeouts.}

\item{years}{Integer range of years from 1950 to 2014. Defaults to 2010}

\item{bouning_wkt}{Well Known Text string representation of a polygon or multipolygon. Calling st_as_text() on an sf
object will produce valid WKT.}
}
\value{
SeaAroundUs formatted data.frame
}
\description{
either download or fetch from cache
}
