% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{flag_lower_outliers}
\alias{flag_lower_outliers}
\title{Flag lower outliers from a single vector}
\usage{
flag_lower_outliers(x, range, m = NULL, s = NULL)
}
\description{
helper function to flag upper outliers from a vector
}
