% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{hotellings_bulk}
\alias{hotellings_bulk}
\title{Hotellings T^2 statistic for many points}
\usage{
hotellings_bulk(means, res_sq)
}
\arguments{
\item{means}{data.frame of sites by predictor, each row is the mean of a sample}

\item{res_sq}{data.frame of sites by diagonal variance, each row is the diagonal variance of a sample}
}
\value{
site by site p-value matrix of all pairs of sites
}
\description{
Calculates the Hotellings T^2 test statistic for all pairs of points at once.
Vectorized for speed.
}
\details{
Unlike hotellings_summary(), hotellings_bulk requires Procrustus residuals to be squared before
being passed in.
}
