% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{hotellings_summary}
\alias{hotellings_summary}
\title{Hotellings T^2 test for a pair of sites}
\usage{
hotellings_summary(x_bar, y_bar, x_var_vec, y_var_vec, x_n = 1,
  y_n = 1)
}
\arguments{
\item{x_bar}{mean of x sample}

\item{y_bar}{mean of y sample}

\item{x_var_vec}{standard deviation of x sample}

\item{y_var_vec}{standard deviation of y sample}

\item{x_n}{number of samples in x}

\item{y_n}{number of samples in y}
}
\value{
list of  \describe{
  \item{\code{df}}{degrees of freedom}
  \item{\code{t2}}{T^2 test statistic}
  \item{\code{f}}{F statistic}
  \item{\code{pval}}{p-value}
 }
}
\description{
Calculates the Hotellings T^2 test statistic and resturns a list of relevant values.
}
\details{
hotellings_summary() assumes that the covariance matricies are both diagonals and
represent standard deviation, not variance.
}
