% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{import_biooracle_env}
\alias{import_biooracle_env}
\title{Import BioORACLE layers from sdmpredictors library}
\usage{
import_biooracle_env(env_vars = c("depth", "temp", "nitrate", "silicate",
  "chlo", "iron", "salinity", "curvel"), env_modes = c("mean", "range"),
  data_dir = "./data")
}
\arguments{
\item{env_vars}{character list of environmental variables. Default is
c("depth", "temp", "nitrate", "silicate", "chlo", "iron",  "salinity", "curvel")}

\item{env_modes}{character list of environmental statistics. Default is c("mean", "range")}

\item{data_dir}{location for sdm_predictors to save downloaded layers into. Defaults to "./data"}
}
\value{
raster brick of all selected environmental layers
}
\description{
Given a list of environmental variables and a list of
statistical measures, returns all valid combinations
of environmental variable and statistical measure.
}
\details{
Currently only sea surface measures are returned. ("ss" only, not "bdmean" "bdmax" or "bdmin")

Available environmental variables are:
"depth", "temp", "nitrate", "silicate", "chlo", "iron",  "salinity", "curvel" (current velocity),
"icethick", "icecover" "dissox" (dissolved oxygen), "phosphate", "lightbot" (light at bottom),
"carbonphyto" (carbon phytoplankton), "pp" (primary production)

Available statistical measures are:
"mean", "range", "min" (average annual min), "max", "ltmin" (long term minimum), "ltmax",
}
