% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{import_seaaroundus_bio}
\alias{import_seaaroundus_bio}
\title{Import SeaAroundUs data}
\usage{

  import_seaaroundus_bio(data_dir = "/vmshare/phd/data/SeaAroundUs/ausEEZcatches.rds",
  sau_groups = c("pelagiclg", "pelagicmd", "pelagicsm"),
  extra_sau_sp = c("Prionace glauca", "Carcharhinus longimanus",
  "Sphyrna lewini", "Sphyrna zygaena", "Alopias vulpinus",
  "Alopias superciliosus", "Lamna nasus", "Emmelichthys nitidus nitidus"),
  years = 2010)
}
\arguments{
\item{data_dir}{Directory to cache SeaAroundUs data}

\item{sau_groups}{Character Vector of SeaAroundUs groups. Defaults to pelagic groups}

\item{extra_sau_sp}{Character Vector of additional SeaAroundUs species. Defaults to pelagic sharks.}

\item{years}{Integer range of years from 1950 to 2014. Defaults to 2010}
}
\description{
Pull in SeaAroundUs data and clean up to site by species matrix
}
\details{
For now, just an RDS file, function is not accessing SeaAroundUs servers or a local cache.
}
