% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{log_env_data}
\alias{log_env_data}
\title{Take log of selected env vars}
\usage{
log_env_data(dataset, exclude_cols = NULL)
}
\arguments{
\item{dataset}{A data.frame to transform}

\item{exclude_cols}{A character vector of col names that will not be transformed}
}
\description{
log_env_data safely applies a log transform to
specific columns in an environmental data.frame
}
\details{
Generally, I want to transform everything except a few key cols
so log_env_data just does that, inverting and offsetting as needed
}
