% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cluster.R
\name{optimal_icl}
\alias{optimal_icl}
\title{Find optimal clusters by according to ICL}
\usage{
optimal_icl(data, mclust_model, n_groups, n_iter = 3, min_site_ratio)
}
\arguments{
\item{data}{A matrix or data.frame to cluster}

\item{mclust_model}{Mclust model string such as "EII". see ?mclustModelNames}

\item{n_groups}{Vector of group counts to test}

\item{n_iter}{Number of times to refit each group count in n_groups}

\item{min_site_ratio}{Only accept fits with all clusters owning at least \code{nrow(data) * min_site_ratio} sites}
}
\value{
list of scores and best mclust object
}
\description{
Find optimal clusters by according to ICL
}
