% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_cluster_map}
\alias{plot_cluster_map}
\title{Plot map coloured by assigned cluster}
\usage{
plot_cluster_map(dataset, col_x = "lon", col_y = "lat",
  depth_col = NULL, depth_contour = NULL, cluster_model,
  legend_thres = 10, sf_poly = NULL)
}
\arguments{
\item{dataset}{A dataframe of observations with an associated fitted model}

\item{col_x}{Column name for x axis}

\item{col_y}{Column name for y axis}

\item{depth_col}{Column name for bathymetry}

\item{depth_contour}{Vector of depths used for contour lines.}

\item{cluster_model}{Cluster model (currently mclust) fitted to \code{dataset}}

\item{legend_thres}{When the number of clusters exceeds legend_thres, the legend is not plotted}

\item{sf_poly}{A polygon of class sf for plotting on the map.}
}
\value{
A ggplot object. The ggplot object is not printed
}
\description{
Produces a map plot coloured by cluster.
A bounding polygon may be provided for context.
If not all columns in dataset were used for fitting the
cluster model, the relevant columns will be read from
the cluster_model
}
