% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_cluster_map_ci}
\alias{plot_cluster_map_ci}
\title{Plot cluster map, but hide sites with low confidence}
\usage{
plot_cluster_map_ci(x, rawx, cluster_model, legend_thres = 10, landPoly,
  confThres)
}
\description{
Plot cluster map, but hide sites with low confidence
}
