% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_cluster_own}
\alias{plot_cluster_own}
\title{Plot clusters individually, coloured by confidence}
\usage{
plot_cluster_own(dataset, col_x = "lon", col_y = "lat",
  depth_col = NULL, depth_contour = NULL, cluster_model,
  sf_poly = NULL)
}
\arguments{
\item{dataset}{A dataframe of observations with an associated fitted model}

\item{col_x}{Column name for x axis}

\item{col_y}{Column name for y axis}

\item{depth_col}{Column name for bathymetry}

\item{depth_contour}{Vector of depths used for contour lines.}

\item{cluster_model}{Cluster model (currently mclust) fitted to \code{dataset}}

\item{sf_poly}{A polygon of class sf for plotting on the map.}
}
\value{
A ggplot object. The ggplot object is not printed
}
\description{
Generates a list of ggplot objects, one per cluster.
}
\details{
Each cluster is plotted alone, showing the sites each cluster is associated with, and how strongly the
site associates with each cluster.

Across all clusters, each site membership sums to 1
}
