% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_cluster_pairs}
\alias{plot_cluster_pairs}
\title{Plot cluster pairs}
\usage{
plot_cluster_pairs(dataset, cluster_model, level = 0.683,
  legend_thres = 10, alpha = 0.3)
}
\arguments{
\item{dataset}{A data.frame of variables to plot.}

\item{cluster_model}{Cluster model (currently mclust) fitted to \code{dataset}}

\item{level}{The heigth of the contour line for the ellipses. Defaults to 68.3%, or 1 standard deviation from the mean}

\item{legend_thres}{When the number of clusters exceeds legend_thres, the legend is not plotted}

\item{alpha}{Alpha value of scatterplot points}
}
\value{
A list of ggplot object. The ggplot objects are not printed
}
\description{
Generates a list of plots, of each pair of variables,
showing clustered data with ellipses.
}
