% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_cor_heatmap}
\alias{plot_cor_heatmap}
\title{Plot correlation heatmap from a data.frame}
\usage{
plot_cor_heatmap(dataset)
}
\arguments{
\item{dataset}{A data.frame to plot}
}
\value{
a ggplot object. The ggplot object is not printed
}
\description{
Plot correlation heatmap from a data.frame
}
