% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_gf_importance}
\alias{plot_gf_importance}
\title{Gradient Forest Variable importance weighted by R^2}
\usage{
plot_gf_importance(x)
}
\arguments{
\item{x}{A gradient forest model}
}
\value{
ggplot2 plot object showing R^2 weighted importance
}
\description{
Plots variable importance weighted by R^2.
}
\details{
Uses S3 dispatch to handle both gradientForest and combinedGradientForest
}
