% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_maps_raster}
\alias{plot_maps_raster}
\title{Plots gridded variables projected onto maps}
\usage{
plot_maps_raster(dataset, lat_col, lon_col, sf_poly = NULL)
}
\arguments{
\item{dataset}{A data.frame of variables to plot.
The data.frame is expected to have either "lat" and "lon" columns, or "x" and "y" columns.}

\item{lat_col}{Name of column containing latitude or y coordinates}

\item{lon_col}{Name of column containing longitude or x coordinates}

\item{sf_poly}{A polygon of class sf for plotting on the map.}
}
\value{
List of ggplot objects. The ggplot objects are not printed
}
\description{
Generates a gglot map for each variable in \code{dataset}.
}
\details{
An sf_polygon can be included to provide context for the data.
}
