% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_marginals}
\alias{plot_marginals}
\title{Plot marginals of a dataset}
\usage{
plot_marginals(dataset, binwidth = 1)
}
\arguments{
\item{dataset}{A data.frame of variables to plot}

\item{binwidth}{Width of bins for the histogram plots. Defaults to 1.}
}
\value{
List of ggplot objects. The ggplot objects are not printed
}
\description{
Returns a list of ggplot objects
showing the marginal histograms of each variable
}
