% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_marginals_species}
\alias{plot_marginals_species}
\title{Plots marginals of species}
\usage{
plot_marginals_species(dataset, binwidth = 1)
}
\arguments{
\item{dataset}{A data.frame of variables to plot}

\item{binwidth}{Width of bins for the histogram plots. Defaults to 1.}
}
\value{
List of ggplot objects. The ggplot objects are not printed
}
\description{
Species observations usually have very large numbers
of 0's, so plot_marginals_species excludes 0 from the
histogram for easier visualisation
}
