% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Plots pairs of variables}
\usage{
plot_pairs(dataset, coord_equal = FALSE)
}
\arguments{
\item{dataset}{A data.frame of variables to plot}
}
\value{
List of ggplot objects. The ggplot objects are not printed
}
\description{
Generates a gglot scatterplot for each pair of variables in \code{dataset}
}
