% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{R.block.multi}
\alias{R.block.multi}
\title{Interpolate residuals}
\usage{
R.block.multi(Xblock, Xsite, res, prox = 1)
}
\description{
Given a set of survey sites with residuals and a set of environmental sites,
estimate the residual at each environmental site with a
Gaussian density kernel.
}
\details{
Distances are in transformed predictor space, not latitude or longitude.
Sites with similar environmental conditions will have similar residuals
regardless of geographic distance.

@param Xblock data frame of site by predictors that need residual estimates
 @param Xsite data.frame of site by predictors that have known residuals
 @param res data.frame of site by predictor_residuals, must match dimensions of Xsite
 @param prox Gaussian kernel scale term. The kernel width is standardised by the median distance between Xsites, prox controls further scaleing of the kernel width.

@return data.frame of residuals for Xblock, rows of the returned data.frame align with rows of Xblock
}
