% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{rare_sp_cols}
\alias{rare_sp_cols}
\title{Filter Rare species}
\usage{
rare_sp_cols(dataset, n = 5, exclude_cols = NULL)
}
\arguments{
\item{dataset}{A data.frame to filter}

\item{n}{minimum number of non-zero observations to accept species}

\item{exclude_cols}{Columns to ignore during filtering, always return true}
}
\value{
boolean vector of columns with rare species
}
\description{
filters out species with fewer than n non-zero observations
}
