% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{sau_id_to_lat_lon}
\alias{sau_id_to_lat_lon}
\title{helper function to convert SeaAroundUs IDs to lat and lon values}
\usage{
sau_id_to_lat_lon(cellId, gridSize = 0.5, centred = TRUE)
}
\description{
helper function to convert SeaAroundUs IDs to lat and lon values
}
