% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cluster.R
\name{withWarnings}
\alias{withWarnings}
\title{Helper function to allow both warnings and values to be captured}
\usage{
withWarnings(expr)
}
\arguments{
\item{expr}{a code block to evaluate}
}
\value{
list of $value and $warnings, $value contains the result of evaluating the code block regardless of warnings,
$warnings contains a list of warnings generated during evaluation
}
\description{
Helper function to allow both warnings and values to be captured
}
