% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{convert_csiro_cpr_copepod}
\alias{convert_csiro_cpr_copepod}
\title{Convert CSIRO CPR copepod CSV into R format}
\usage{

  convert_csiro_cpr_copepod(in_file = "/vmshare/phd/data/CSIRO_cpr/cpr_copepods_matrix.csv",
  date_format = "ymd HMS")
}
\arguments{
\item{in_file}{Pathe to the original CSV file. Default "/vmshare/phd/data/CSIRO_cpr/cpr_copepods_matrix.csv"}

\item{date_format}{lubridate::parse_date_time() orders string for converting dates into standard R format. Defaults to "ymd HMS"}
}
\value{
A list containing: \describe{
  \item{\code{data}}{A site by species data_frame. Metadata columns "lat", "lon" and "date",
  and one column per species.}
  \item{\code{sp_names}}{Names of species columns in \code{data}}
 }
}
\description{
Takes the CSIRO CPR copepod data in CSV form and converts it into
a form suitable for processing with Gradient Forest, then
exports to different formats. The species list is exported
to the same format.
}
\details{
The magic numbers in this function are:
column indexes and names for non-Species columns
}
