% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{gf_plots}
\alias{gf_plots}
\title{GradientForest summary plot collection}
\usage{
gf_plots(gf_fit, vars)
}
\description{
Generates a set of gradientForest plots
}
\details{
These are base R plots, apart from the variable importance plot
}
