% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{import_seaaroundus_bio}
\alias{import_seaaroundus_bio}
\title{Import SeaAroundUs data}
\usage{

  import_seaaroundus_bio(data_dir = "/vmshare/phd/data/SeaAroundUs/ausEEZcatches.rds",
  sau_groups = c("pelagiclg", "pelagicmd", "pelagicsm"),
  extra_sau_sp = c("Prionace glauca", "Carcharhinus longimanus",
  "Sphyrna lewini", "Sphyrna zygaena", "Alopias vulpinus",
  "Alopias superciliosus", "Lamna nasus", "Emmelichthys nitidus nitidus"),
  bounding_wkt = NULL, chunk_size = 100, years = 2010)
}
\arguments{
\item{data_dir}{Directory to cache SeaAroundUs data}

\item{sau_groups}{Character Vector of SeaAroundUs groups. Defaults to pelagic groups}

\item{extra_sau_sp}{Character Vector of additional SeaAroundUs species. Defaults to pelagic sharks.}

\item{chunk_size}{Integer indicating how many cells to fetch in a single call. Too small will slow down the download process, too large will cause server timeouts.}

\item{years}{Integer range of years from 1950 to 2014. Defaults to 2010}

\item{bouning_wkt}{Well Known Text string representation of a polygon or multipolygon. Calling st_as_text() on an sf
object will produce valid WKT.}
}
\description{
Pull in SeaAroundUs data and clean up to site by species matrix
}
\details{
Only samples resolved down to the species level are included.
}
