% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmethods-package.R
\docType{package}
\name{rmethods-package}
\alias{rmethods}
\alias{rmethods-package}
\title{rmethods: Methods for Phil Dyers PhD}
\description{
Set of functions used in my PhD.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PhDyellow/rmethods}
  \item Report bugs at \url{https://github.com/PhDyellow/rmethods/issues}
}

}
\author{
\strong{Maintainer}: Philip Dyer \email{phildyer@protonmail.com}

}
\keyword{internal}
